add confirmation on nwb, snakemake params change
