+++ frontend/src/components/Workspace/FlowChart/ReactFlowComponent.tsx
@@ -27,17 +27,6 @@ import {
 import "reactflow/dist/style.css"
 
 import "style/flow.css"
-import { FormHelperText, Popover } from "@mui/material"
-
-import { AlgorithmOutputDialog } from "components/Workspace/FlowChart/Dialog/AlgorithmOutputDialog"
-import { ClearWorkflowIdDialog } from "components/Workspace/FlowChart/Dialog/ClearWorkflowIdDialog"
-import {
-  ClearWorkflowIdDialogValue,
-  DialogContext,
-  ErrorDialogValue,
-  FileSelectDialogValue,
-} from "components/Workspace/FlowChart/Dialog/DialogContext"
-import { FileSelectDialog } from "components/Workspace/FlowChart/Dialog/FileSelectDialog"
 import {
   DND_ITEM_TYPE_SET,
   TreeItemCollectedProps,
@@ -63,21 +52,6 @@ import {
 } from "store/slice/FlowElement/FlowElementSlice"
 import { NodeData } from "store/slice/FlowElement/FlowElementType"
 import { UseRunPipelineReturnType } from "store/slice/Pipeline/PipelineHook"
-import { clearCurrentPipeline } from "store/slice/Pipeline/PipelineSlice"
-
-const initDialogFile = {
-  filePath: "",
-  open: false,
-  fileTreeType: undefined,
-  multiSelect: false,
-  onSelectFile: () => null,
-}
-
-const initClearWorkflow = {
-  open: false,
-  handleOk: () => null,
-  handleCancel: () => null,
-}
 
 const ReactFlowProviderComponent = ReactFlowProvider as FC<{
   children: ReactNode
@@ -90,15 +64,6 @@ export const ReactFlowComponent = memo(function ReactFlowComponent(
   const flowEdges = useSelector(selectFlowEdges)
   const egdes = flowEdges.filter((item) => !isNode(item)) as Edge<NodeData>[]
   const dispatch = useDispatch()
-  const [dialogNodeId, setDialogNodeId] = useState("")
-  const [dialogFile, setDialogFile] =
-    useState<FileSelectDialogValue>(initDialogFile)
-  const [dialogClearWorkflowId, setDialogClearWorkflowId] =
-    useState<ClearWorkflowIdDialogValue>(initClearWorkflow)
-  const [messageError, setMessageError] = useState<ErrorDialogValue>({
-    anchorElRef: { current: null },
-    message: "",
-  })
 
   const onConnect = (params: Connection | Edge) => {
     dispatch(
@@ -177,103 +142,32 @@ export const ReactFlowComponent = memo(function ReactFlowComponent(
   )
   return (
     <div className="flow">
-      <DialogContext.Provider
-        value={{
-          onOpenOutputDialog: setDialogNodeId,
-          onOpenFileSelectDialog: setDialogFile,
-          onOpenClearWorkflowIdDialog: setDialogClearWorkflowId,
-          onMessageError: setMessageError,
-        }}
-      >
-        <ReactFlowProviderComponent>
-          <div className="reactflow-wrapper" ref={wrapparRef}>
-            <ReactFlow
-              ref={drop}
-              nodes={flowNodes}
-              edges={flowEdges}
-              onNodesChange={onNodesChange}
-              onEdgesChange={onEdgesChange}
-              onConnect={onConnect}
-              onInit={onInit}
-              onDragOver={onDragOver}
-              onNodeDragStop={onNodeDragStop}
-              nodeTypes={reactFlowNodeTypes}
-              edgeTypes={reactFlowEdgeTypes}
-              defaultViewport={{
-                x: flowPosition[0],
-                y: flowPosition[1],
-                zoom: flowPosition[2],
-              }}
-              onMoveEnd={onMoveEnd}
-            >
-              <ToolBar {...props} />
-              <Controls />
-            </ReactFlow>
-          </div>
-        </ReactFlowProviderComponent>
-        {dialogNodeId && (
-          <AlgorithmOutputDialog
-            nodeId={dialogNodeId}
-            open
-            onClose={() => setDialogNodeId("")}
-          />
-        )}
-        {dialogFile.open && (
-          <FileSelectDialog
-            multiSelect={dialogFile.multiSelect}
-            initialFilePath={dialogFile.filePath}
-            open={dialogFile.open}
-            onClickOk={(path: string | string[]) => {
-              dialogFile.onSelectFile(path)
-              setDialogFile(initDialogFile)
+      <ReactFlowProviderComponent>
+        <div className="reactflow-wrapper" ref={wrapparRef}>
+          <ReactFlow
+            ref={drop}
+            nodes={flowNodes}
+            edges={flowEdges}
+            onNodesChange={onNodesChange}
+            onEdgesChange={onEdgesChange}
+            onConnect={onConnect}
+            onInit={onInit}
+            onDragOver={onDragOver}
+            onNodeDragStop={onNodeDragStop}
+            nodeTypes={reactFlowNodeTypes}
+            edgeTypes={reactFlowEdgeTypes}
+            defaultViewport={{
+              x: flowPosition[0],
+              y: flowPosition[1],
+              zoom: flowPosition[2],
             }}
-            onClickCancel={() => {
-              setDialogFile(initDialogFile)
-            }}
-            fileType={dialogFile.fileTreeType}
-          />
-        )}
-        {dialogClearWorkflowId.open && (
-          <ClearWorkflowIdDialog
-            open={dialogClearWorkflowId.open}
-            handleOk={() => {
-              dispatch(clearCurrentPipeline())
-              dialogClearWorkflowId.handleOk()
-              setDialogClearWorkflowId(initClearWorkflow)
-            }}
-            handleCancel={() => {
-              dialogClearWorkflowId.handleCancel()
-              setDialogClearWorkflowId(initClearWorkflow)
-            }}
-          />
-        )}
-        {messageError?.message && (
-          <Popover
-            open
-            anchorEl={messageError.anchorElRef.current}
-            onClose={() =>
-              setMessageError({
-                anchorElRef: { current: null },
-                message: "",
-              })
-            }
-            anchorOrigin={{
-              vertical: "top",
-              horizontal: "right",
-            }}
-            transformOrigin={{
-              vertical: "bottom",
-              horizontal: "left",
-            }}
+            onMoveEnd={onMoveEnd}
           >
-            <div style={{ margin: 8 }}>
-              <FormHelperText error={true}>
-                {messageError.message}
-              </FormHelperText>
-            </div>
-          </Popover>
-        )}
-      </DialogContext.Provider>
+            <ToolBar {...props} />
+            <Controls />
+          </ReactFlow>
+        </div>
+      </ReactFlowProviderComponent>
     </div>
   )
 })
