+++ frontend/src/components/Workspace/FlowChart/RightDrawer/SnakemakeSettingContents.tsx
@@ -41,6 +41,7 @@ const ParamItem = memo(function ParamItem({ paramKey }: ParamItemProps) {
     paramValueSelector: selectSnakemakeParamsValue,
     paramUpdateActionCreator: (path, newValue) =>
       updateParam({ path, newValue }),
+    requireConfirm: true,
   })
   return <Component paramKey={paramKey} />
 })
